+++ api/src/main/java/com/projectspring/api/Repositories/PlaceSubTypeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.projectspring.api.Entities.PlaceSubType;
 
-public interface PlaceSubTypeRepository extends JpaRepository<PlaceSubType, Integer> {
+public interface PlaceSubTypeRepository extends JpaRepository<PlaceSubType, Long> {
     
 }
